Ignore Container Workspace Directory, Install a Ruby version for RVM
Container Workspace directory can be ignored because default behaviour of repo is to create the directory, if not present. This is where the Rails app should reside once in the environment. At that point the Rails app can become its own repo. Dockerfile changes result in successfully configuring RVM to use a Ruby version.